Rename theano.link.c.vm to theano.link.vm
The `CVM` class definition was also split off into `theano.link.c.cvm`.
